fixed form helper not honoring div attributes
Input options can have a div attribute that according to the latest cakebook can be either a string of classes, or an array of HTML attributes for the div. TwitterBootstrapFormHelper was incorrectly passing an array to Helper::addClass which issues a warning when trimming the string.

I changed it to just pass the div options array to Helper::addClass instead of the whole options array, and changed the rest of the class to use the array form instead of the string form for the div option.